+++ src/python.cpp
@@ -2694,38 +2694,38 @@ SEXP py_run_string_impl(const std::string& code,
                         bool local = false,
                         bool convert = true)
 {
-  // get main module. note that PyModule_GetDict returns
-  // a borrowed reference and so if we're returning a reference
-  // to that dictionary (with local = false) then we need to
-  // Py_IncRef it
+  // retrieve reference to main module dictionary
+  // note: both PyImport_AddModule() and PyModule_GetDict()
+  // return borrowed references
   PyObject* main = PyImport_AddModule("__main__");
-  PyObject* main_dict = PyModule_GetDict(main);
-  
-  // use separate dictionary for locals if requested
-  PyObject* local_dict = NULL;
+  PyObject* globals = PyModule_GetDict(main);
   
   if (local) {
-    PyObjectPtr local_dict_ptr(PyDict_New());
-    local_dict = local_dict_ptr.get();
+    
+    // create dictionary to capture locals
+    PyObjectPtr locals(PyDict_New());
+    
+    // run the requested code
+    PyObjectPtr res(PyRun_StringFlags(code.c_str(), Py_file_input, globals, locals, NULL)); 
+    if (res.is_null())
+      stop(py_fetch_error());
+    
+    // return locals dictionary (detach so we don't decref on scope exit)
+    return py_ref(locals.detach(), convert);
+    
   } else {
-    local_dict = main_dict;
+    
+    // run the requested code
+    PyObjectPtr res(PyRun_StringFlags(code.c_str(), Py_file_input, globals, globals, NULL)); 
+    if (res.is_null())
+      stop(py_fetch_error());
+    
+    // because globals is borrowed, we need to incref here
+    Py_IncRef(globals);
+    return py_ref(globals, convert);
+    
   }
   
-  // run the requested code
-  PyObjectPtr res(
-      PyRun_StringFlags(
-        code.c_str(),
-        Py_file_input,
-        main_dict,
-        local_dict,
-        NULL));
-  
-  if (res.is_null())
-    stop(py_fetch_error());
-
-  // return dictionary with objects defined during the execution
-  Py_IncRef(local_dict);
-  return py_ref(local_dict, convert);
 }